Move the ic_error png to res/drawable, from res/drawable-nodpi
When using a device that is >= API 21, the vector drawable error icon
in the `drawable-v21` resource directory is not used - the png in the
`drawable-nodpi` is used instead.

I would need to add the `nodpi` qualifier to ensure the vector drawable
is used when applicable. I am moving the png icon to the no-qualifier
drawable directory - not amending 'drawable-v21' - to remain
consistent (I am storing other icons there).