Simplify logic of updating position(s) when rearranged
I was previously incrementing/decrement the positions of effected
Groups/Items by evaluating the difference between the old and new
positions. The code was brittle and hard to understand.

I now update the moved Group/Item to its new position (and add 0.5 to
ensure it is above the Group/Item that occupies that position) then
update all appropriate Groups/Items to ensure all are in consecutive
order and integers.