Fix locally initiated changes being pushed to the UI twice in-a-row
When a locally initiated change occurs it is pushed to the UI twice:
1) The local cache is updated, thus the EventListener is notified and
the changes are pushed to the UI.
2) When the remote database is updated, it returns with the updated data
and EventListener is notified again.

I added a flag to keep track of whether or not a local change was
recently made - if so, only push the first change from the local cache,
not the second change from the remote database.