Fix animation lag when dragging a list item in RecyclerView
The lag was caused by the background behind the list item that would
show when the list item is swiped.

Calculating the size of the list item (where to draw the red background)
then calculating the placement of the delete icon was a resource
intensive process. Moreover, those calculation would happen for every
single list item.

The solution (thanks to **AndroidHive**, linked below) is to give each
list item two layers. The foreground is the normal list item, the
background is red with a delete icon. When swiped, only the foreground
moves and the background is exposed to the user.

https://www.androidhive.info/2017/09/android-recyclerview-swipe-delete-undo-using-itemtouchhelper/