Remove creation of temp ArrayList prior to deletion of Groups/Items
I was copying the List of temp Groups/Items to another List because I
use to use RxJava to move the deletion to the I/O thread. This ensured
the List of Groups/Items to be deleted was still there when the delete
method was invoked - the ViewModel temp List is cleared at the last
line of the method.