Remove minify from the debug build
I enabled minify for the debug build because I reached Dex's 64k limit
after adding Firebase dependencies months ago. I am now removing it
because it is no longer needed and it prevents Instant Run from working
(among other issues).

I do not know why it is no longer needed, as I have not removed any
dependencies. I do know that it is not because of a newer version of
Gradle nor Android's plugin - I rollback to the versions I was using
when I first added minify to the debug build.