Removed observing of repo from AddEditDialog
UserList/Item will be refereed to as "item".

If a item is added in the background while a user is adding a item,
it will be added one-above the item added in the background. That is
because the new item's position is static - its passed-in when the Dialog
is created.

I refactored to this because observing the repo was overkill for simply
adding an item. Besides, it is OK if two different items share the same
position.