Added button to check email verified
Normally, the user would press the Continue button on the web page
that confirms that their email has been verified - that would
automatically re-open the app and check if their email has, if fact,
been verified.

If the user, instead, manually comes back to the app after that web page,
this button will let the user manually initiate the checking process.