Refactored EmailReAuth
Pressing back and clicking the cancel button bought the user to different
locations - the former closed the app, while the latter bought the user
to the the main view.

Now, when the user presses back, or navigates up, they return to the
PreferencesView. A uniform destination provides a superior UX.

Additionally, I removed the cancel button from the UI. This mirrors how
the initial authentication operates. There is no cancel button, the user
simply presses back to cancel.