Fixed exception that occurred when user selected a UserList
When a UserList is moved to the top of the list, its position is
temporally -1 while the positions of all UserLists are being updated.
If the user selects that UserList during that period of time, an
ArrayIndexOutOfBoundsException is thrown because -1 is an invalid
position. This fix ensures that invalid position is caught.
The user will merely assume their click did not register and try again.