Removed JUnit resource that set the test lifecycle to per_class
I originally added it to reduce the amount of time it takes to execute
all of the unit tests. At the time, there was no foreseeable down side
to doing so. The down side causing me to remove this is the current
refactoring to an event-based solution that relies upon LiveData. To
ensure old data in the LiveData observable is clear prior to the next
test, I need to manually re-instantiate the class under test, thus
negating the change to JUnit's test lifecycle.