Fixed the Gradle assemble command failing
If a debug or release APK is built via Android Studio's GUI, it builds without an issue. If an APK is built via the command line, specifically the `assemble` command, the build fails stating that it cannot link a resource  file in the data module.

The issue has to do with the Firebase UI dependency. The compiler cannot find the aforementioned resource file because the data module does not contain the design library resources required by the Firebase UI dependency.